[8496] Get archetypes from ui config in useArchetypes hook
