+++ ui/app/src/state/reducers/uiConfig.ts
@@ -25,10 +25,8 @@ import {
 import { changeSiteComplete } from '../actions/sites';
 import { fetchSiteLocales, fetchSiteLocalesComplete, fetchSiteLocalesFailed } from '../actions/translation';
 import { deserialize, fromString, serialize } from '../../utils/xml';
-import { applyDeserializedXMLTransforms, extendArchetypeDescriptor } from '../../utils/object';
+import { applyDeserializedXMLTransforms } from '../../utils/object';
 import { getUserLocaleCode, getUserTimeZone } from '../../utils/datetime';
-import type { ContentType } from '../../models';
-import { asArray } from '../../utils/array';
 
 const initialState: GlobalState['uiConfig'] = {
 	error: null,
@@ -57,7 +55,6 @@ const initialState: GlobalState['uiConfig'] = {
 		}
 	},
 	references: null,
-	archetypes: null,
 	xml: null,
 	publishing: {
 		deleteCommentRequired: false,
@@ -83,7 +80,6 @@ const reducer = createReducer<GlobalState['uiConfig']>(initialState, (builder) =
 		.addCase(fetchSiteUiConfigComplete, (state, { payload }) => {
 			let config = payload.config;
 			const references = {};
-			const archetypes = {};
 			if (config) {
 				const configDOM = fromString(config);
 				const site = payload.site;
@@ -108,40 +104,6 @@ const reducer = createReducer<GlobalState['uiConfig']>(initialState, (builder) =
 
 				configDOM.querySelectorAll('widget').forEach((e, index) => e.setAttribute('uiKey', String(index)));
 
-				configDOM
-					.querySelectorAll('[id="craftercms.components.ContentTypeManagement"] > configuration > objectTypes')
-					.forEach((tag) => {
-						const descriptor = tag.querySelector('descriptor');
-						// Parent archetypes must be defined before children so that they can be extended properly
-						const extendsFrom = tag.getAttribute('extends');
-						let parentArchetype = null;
-						if (extendsFrom) {
-							parentArchetype = archetypes[extendsFrom] ?? {};
-						}
-						archetypes[tag.id] = {
-							...parentArchetype,
-							id: tag.id
-						};
-						if (descriptor) {
-							let deserializedDescriptor: ContentType = deserialize(descriptor.innerHTML);
-							deserializedDescriptor = {
-								...deserializedDescriptor,
-								id: tag.id,
-								sections: asArray(deserializedDescriptor.sections).map((section) => ({
-									...section,
-									fields: asArray(section.fields)
-								})),
-								dataSources: asArray(deserializedDescriptor.dataSources)
-							};
-
-							archetypes[tag.id] = {
-								...archetypes[tag.id],
-								name: deserializedDescriptor.name,
-								descriptor: extendArchetypeDescriptor(parentArchetype?.descriptor, deserializedDescriptor)
-							};
-						}
-					});
-
 				config = serialize(configDOM);
 			}
 
@@ -149,8 +111,7 @@ const reducer = createReducer<GlobalState['uiConfig']>(initialState, (builder) =
 				...state,
 				isFetching: false,
 				xml: config,
-				references,
-				archetypes
+				references
 			};
 		})
 		.addCase(fetchSiteUiConfigFailed, (state, { payload }) => ({
